services/nature-forge-ui/requirements.txt
fastapi[all]
uvicorn
httpx
Pillow
trimesh
fast-simplification
